+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Wed, 25 Aug 2021 13:56:22 GMT and should not be manually modified.
+This log was last generated on Thu, 26 Aug 2021 17:52:10 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.0.0-beta.33
+
+Thu, 26 Aug 2021 17:52:10 GMT
+
+### Changes
+
+- Bump @ni/nimble-angular to v1.0.0-beta.33 (1458528+fredvisser@users.noreply.github.com)
+
+## 1.0.0-beta.32
+
+Thu, 26 Aug 2021 15:36:09 GMT
+
+### Changes
+
+- Bump @ni/nimble-angular to v1.0.0-beta.32 (rajsite@users.noreply.github.com)
 
 ## 1.0.0-beta.31